moving test to manifest_generator_test.go
